User create and delete updated

+++ AccountService/Controllers/UserController.cs
@@ -4,6 +4,7 @@ using AccountService.Handlers.Address;
 using AccountService.Handlers.User;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
+using AccountService.Handlers.Links;
 
 namespace AccountService.Controllers
 {
@@ -15,22 +16,41 @@ namespace AccountService.Controllers
 
         private readonly IUserHandler userHandler;
         private readonly IAddressHandler addressHandler;
+        private readonly IOrderLink orderLink;
+        private readonly IReservationLink reservationLink;
 
 
-        public UserController(IUserHandler userHandler, IAddressHandler addressHandler)
+
+        public UserController(IUserHandler userHandler, 
+                                IAddressHandler addressHandler,
+                                IOrderLink orderLink,
+                                IReservationLink reservationLink)
         {
             this.userHandler = userHandler;
             this.addressHandler = addressHandler;
+            this.orderLink = orderLink;
+            this.reservationLink = reservationLink;
         }
 
         [HttpPost]
         [AllowAnonymous]
         public async Task<ActionResult<UserDTO>> CreateUser([FromBody] CreateUserDTO createUserDTO)
         {
-            var address = await addressHandler.CreateAddress(createUserDTO.Address);
-            var user = await userHandler.CreateUser(createUserDTO, address.IdAddress);
-            user.Address = address;
-            return Ok(user);
+            AddressDTO address = new AddressDTO();
+            UserDTO user;
+            if (createUserDTO.Address != null)
+            {
+                address = await addressHandler.CreateAddress(createUserDTO.Address);
+                user = await userHandler.CreateUserWithAddress(createUserDTO, address.IdAddress);
+
+                user.Address = address;
+            }
+            else
+            {
+                user = await userHandler.CreateUser(createUserDTO);
+            }
+
+                return Ok(user);
         }
 
         [HttpGet]
@@ -52,8 +72,12 @@ namespace AccountService.Controllers
         public async Task<ActionResult<UserDTO>> UpdateUser([FromRoute] int idUser, [FromBody] UpdateUserDTO updateUserDTO)
         {
             var user = await userHandler.UpdateUser(updateUserDTO,idUser);
-            var address = await addressHandler.UpdateAddress(updateUserDTO.Address, user.Address.IdAddress);
-            user.Address = address;
+            if(updateUserDTO.Address != null)
+            {
+                var address = await addressHandler.UpdateAddress(updateUserDTO.Address, user.Address.IdAddress);
+                user.Address = address;
+            }
+
             return Ok(user);
         }
 
@@ -66,9 +90,23 @@ namespace AccountService.Controllers
 
 
         [HttpDelete("{idUser}")]
-        public async Task<ActionResult<UserDTO>> DeleteUser([FromRoute] int idUser)
+        public async Task<ActionResult<UserDTO>> DeleteUser([FromRoute] int idUser, [FromHeader] string? authorization)
         {
-            var user = await userHandler.DeleteUser(idUser);
+            var user = await userHandler.GetUserById(idUser);
+            bool? isInUseInReservations = await reservationLink.IsUserInUse(idUser, authorization);
+            bool? isInUseInOrders = await orderLink.IsUserInUse(idUser, authorization);
+
+            if(isInUseInReservations.HasValue == true)
+            {
+                return Conflict("User has reservations.");
+            }
+            else if (isInUseInOrders.HasValue == true)
+            {
+                return Conflict("User has orders.");
+
+            }
+
+            var isDeleted = await userHandler.DeleteUser(idUser);
             return Ok(user);
         }
     }
